# Conflicts:
#	bank-modules/bank-manager/src/main/java/com/zaomengjia/bankmanager/SeckillSystemAdminApplication.java
#	bank-modules/bank-manager/src/main/java/com/zaomengjia/bankmanager/controller/UserController.java
#	bank-modules/bank-manager/src/main/resources/application.yml

+++ bank-modules/bank-manager/src/main/java/com/zaomengjia/bankmanager/service/UserService.java
@@ -50,7 +50,7 @@ public interface UserService {
      * @param id
      * @return
      */
-    int deleteUser(int id);
+    int deleteUser(long id);
 
     /**
      * 更新管理员、用户
